mobile/android/tests/browser/chrome/test_reader_view.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1158885
Migrated from Robocop: https://bugzilla.mozilla.org/show_bug.cgi?id=1184186
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1158885</title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SpawnTask.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://global/skin"/>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"/>
  <script type="application/javascript" src="head.js"></script>
  <script type="application/javascript;version=1.7">

  const { classes: Cc, interfaces: Ci, utils: Cu } = Components;

  Cu.import("resource://gre/modules/Services.jsm");

  add_task(function* test_reader_view_visibility() {
    let gWin = Services.wm.getMostRecentWindow("navigator:browser");
    let BrowserApp = gWin.BrowserApp;

    let url = "http://mochi.test:8888/chrome/mobile/android/tests/browser/chrome/basic_article.html";
    let browser = BrowserApp.addTab("about:reader?url=" + url).browser;

    SimpleTest.registerCleanupFunction(function() {
      BrowserApp.closeTab(BrowserApp.getTabForBrowser(browser));
    });

    yield promiseBrowserEvent(browser, "load");

    let doc = browser.contentDocument;
    let title = doc.getElementById("reader-title");

    // We need to wait for reader content to appear because AboutReader.jsm
    // asynchronously fetches the content after about:reader loads.
    yield promiseNotification("AboutReader:Ready");
    is(title.textContent, "Article title", "found expected content");
  });

  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1158885">Mozilla Bug 1158885</a>
<br>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1184186">Migrated from Robocop testReaderView</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
